♻️ refactor: split credentials module into subpackage
Extract credentials.py into credentials/ with separate files for the
protocol, file backend, vaultwarden backend, and registry. Public API
unchanged via __init__.py re-exports.

Made-with: Cursor
